Encapsulate detail::msghub_impl as msghub::impl
This gives the compiler more opportunities to inline/optimize
